@@ -1,28 +1,31 @@
 {
-  "name": "flee",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "author": "Michael Depp",
-  "license": "ISC",
-  "dependencies": {
-    "@chakra-ui/core": "^1.0.0-rc.4",
-    "@chakra-ui/icons": "^1.0.0-next.3",
-    "firebase": "^8.0.2",
-    "hamburger-react": "^2.2.0",
-    "next": "^10.0.1",
-    "next-pwa": "^3.1.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-icons": "^3.11.0",
-    "showdown": "^1.9.1",
-    "styled-components": "^5.2.0",
-    "styled-loaders-react": "^1.1.0",
-    "styled-theming": "^2.2.0"
-  }
+	"name": "flee",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start"
+	},
+	"author": "Michael Depp",
+	"license": "ISC",
+	"dependencies": {
+		"@chakra-ui/react": "^2.10.2",
+		"@chakra-ui/icons": "^2.1.1",
+		"@emotion/react": "^11.13.3",
+		"@emotion/styled": "^11.13.0",
+		"framer-motion": "^11.3.19",
+		"firebase": "^11.0.2",
+		"hamburger-react": "^2.5.2",
+		"next": "^15.0.3",
+		"next-pwa": "^5.6.0",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
+		"react-icons": "^5.3.0",
+		"showdown": "^2.1.0",
+		"styled-components": "^6.1.13",
+		"react-spinners": "^0.15.0",
+		"styled-theming": "^2.2.0"
+	}
 }
