@@ -1,9 +1,9 @@
-import { CubeGrid } from "styled-loaders-react";
+import { GridLoader } from "react-spinners";
 
 const Spinner = () => (
   <div className="loader">
     <div className="cube">
-      <CubeGrid color="#304FFF" />
+      <GridLoader color="#304FFF" />
     </div>
   </div>
 );
