@@ -1,12 +1,13 @@
-const withPWA = require("next-pwa");
-const prod = process.env.NODE_ENV === "production";
-
-module.exports = withPWA({
-  pwa: {
-    disable: prod ? false : true,
-    dest: "public",
-    register: true,
-    scope: "/",
-    sw: "service-worker.js",
-  },
+const withPWA = require('next-pwa')({
+	dest: 'public',
+	register: true,
+	skipWaiting: true,
+	disable: process.env.NODE_ENV === 'development',
 });
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+	reactStrictMode: true,
+};
+
+module.exports = withPWA(nextConfig);
